32 bit word being passed as an int
git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@6823 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/lib/krb5/asn.1/asn1_k_encode.c
@@ -778,7 +778,7 @@ asn1_error_code asn1_encode_etype_info_entry(buf, val, retlen)
      return ASN1_MISSING_FIELD;
 
   if (val->length)
-	  asn1_addlenfield(val->length,val->salt,1,
+	  asn1_addlenfield((int) val->length,val->salt,1,
 			   asn1_encode_octetstring);
   asn1_addfield(val->etype,0,asn1_encode_integer);
   asn1_makeseq();
